Fix interface include directory, add C++ wrapper
Fix the interface includes directory for the LCM types library; it
should be `include/lcmtypes`, not just `include`, since the headers
themselves may refer to other headers without the `lcmtypes/` prefix.
Also add an INTERFACE library for the C++ bindings. (Since we don't have
internal consumers, this is just an overly glorified mechanism to
propagate an interface include directory.)